Split cache out to own class
