+++ osu.Framework/Graphics/OpenGL/Buffers/FrameBuffer.cs
@@ -4,58 +4,13 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
 using osu.Framework.Graphics.OpenGL.Textures;
 using osu.Framework.Graphics.Textures;
-using osu.Framework.Statistics;
 using osuTK;
 using osuTK.Graphics.ES30;
 
 namespace osu.Framework.Graphics.OpenGL.Buffers
 {
-    internal static class FrameBufferTextureCache
-    {
-        private static readonly List<TextureGLSingle> available_textures = new List<TextureGLSingle>();
-
-        public static TextureGLSingle Get(int width, int height, All filteringMode = All.Linear)
-        {
-            lock (available_textures)
-            {
-                var tex = available_textures.FirstOrDefault(t => t.Width >= width && t.Height >= height && t.FilteringMode == filteringMode) ?? available_textures.FirstOrDefault();
-
-                GlobalStatistics.Get<int>("Native", "FrameBuffer Cache InUse").Value++;
-
-                if (tex != null)
-                {
-                    available_textures.Remove(tex);
-                    return tex;
-                }
-
-                GlobalStatistics.Get<int>("Native", "FrameBuffer Cache Total").Value++;
-                return new FrameBufferTexture(width, height, filteringMode);
-            }
-        }
-
-        public static void Return(TextureGLSingle texture)
-        {
-            lock (available_textures)
-            {
-                available_textures.Add(texture);
-                GlobalStatistics.Get<int>("Native", "FrameBuffer Cache InUse").Value--;
-            }
-        }
-
-        private class FrameBufferTexture : TextureGLSingle
-        {
-            public FrameBufferTexture(int width, int height, All filteringMode = All.Linear)
-                : base(width, height, true, filteringMode)
-            {
-                SetData(new TextureUpload());
-                Upload();
-            }
-        }
-    }
-
     public class FrameBuffer : IDisposable
     {
         private int frameBuffer = -1;
